Fix run_query ommiting last result
The problem is in group() funciton of the parser that groups same results of a time range into a single reuslt, in which the time range is added to group dict only after a new value, not in the current range is detected.
The fix is to add an extra check of range after the loop that creates the ranges finishes.