Fix raw tuple query bug by wrapping root Raw operator
A raw query containing a tuple such as "x,x + 1" was parsed together and couldn't be evaled in the main parse method of PaladinNativeParser, now if the root query is raw, its also wrapped with a lambda var, therfore the evaluation of such query is simply replacing the lambda var. E.g, for "x, x+1" the parser's operators dict is: {__O0: Raw("x, x+1")} and the parsed query is "__O0".