Find a bug in graphs.py debugged with PaLaDiN
The bug was in shortes_path in which `for neighbor in neighbors.union(unvisited)` is incorrect, and should be: `for neighbor in neighbors.intersection(unvisited)`.
Also `current` Vertex should have been removed from distances after it had been added to prev so current is not "stuck" on the same vertex.

This bug was found by retrieving the values of "e, current, neighbors, distances".